HLT-compatible LHCInfoPer* PopCon (duringFill mode)
adjust LHCInfoPer*PopConSourceHandler duringFill mode for HLT-compatibility

fix LHCInfoPerLS PopCon duringFill startSampleTime error

fix filtering in  duringFill mode (prevPayload error)

Remove adding empty payload to empty tag

Extract SourceHandler from LHCInfoPer*PopConAnalyzer

Make perLS and perFill duringFill use OnlinePopCon

Attempt to configure OnlineDBOutputService in LHCInfoPer*PopConAnalyzer_cfg.py

finish LHCInfoPer*PopConAnalyzer_cfg files for updated duringFill

For perLS duringfill implement adding defaultPayload and skipping invalid payloads

implement adding defaultPayload in duringFill mode

duringFill perLS: skip upload if invalid values

Adjust duringFill unit tests

remove duplicated kLumisectionsQueryLimit

fix unit tests, print logs of failed lhcInfoNewPopCon test

add connection strings as cfg parameters, fix frontier key parameter for perFill
