+++ CondTools/RunInfo/plugins/BuildFile.xml
@@ -59,6 +59,10 @@
   <flags EDM_PLUGIN="1"/>
 </library>
 
+<library file="LHCInfoPerLSWriter.cc" name="LHCInfoPerLSWriter">
+  <flags EDM_PLUGIN="1"/>
+</library>
+
 <library file="LHCInfoPerFillAnalyzer.cc" name="LHCInfoPerFillAnalyzer">
   <flags EDM_PLUGIN="1"/>
 </library>
@@ -67,10 +71,14 @@
   <flags EDM_PLUGIN="1"/>
 </library>
 
+<library file="LHCInfoPerLSOnlinePopConAnalyzer.cc" name="CondToolsLHCInfoPerLSOnlinePopConAnalyzer">
+  <flags EDM_PLUGIN="1"/>
+</library>
+
 <library file="LHCInfoPerFillPopConAnalyzer.cc" name="CondToolsLHCInfoPerFillPopConAnalyzer">
   <flags EDM_PLUGIN="1"/>
 </library>
 
-<library file="LHCInfoPerLSWriter.cc" name="LHCInfoPerLSWriter">
+<library file="LHCInfoPerFillOnlinePopConAnalyzer.cc" name="CondToolsLHCInfoPerFillOnlinePopConAnalyzer">
   <flags EDM_PLUGIN="1"/>
 </library>
