+++ CondTools/RunInfo/python/LHCInfoPerLSPopConAnalyzer_cfg.py
@@ -3,10 +3,6 @@ import FWCore.ParameterSet.Config as cms
 import FWCore.ParameterSet.VarParsing as VarParsing
 process = cms.Process("LHCInfoPerLSPopulator")
 from CondCore.CondDB.CondDB_cfi import *
-#process.load("CondCore.DBCommon.CondDBCommon_cfi")
-#process.CondDBCommon.connect = 'sqlite_file:lhcinfoperls_pop_test.db'
-#process.CondDBCommon.DBParameters.authenticationPath = '.'
-#process.CondDBCommon.DBParameters.messageLevel=cms.untracked.int32(1)
 
 sourceConnection = 'oracle://cms_omds_adg/CMS_RUNINFO_R'
 if socket.getfqdn().find('.cms') != -1:
@@ -60,6 +56,109 @@ options.register( 'endTime'
                      processes only fills starting before endTime;
                      default to empty string which sets no restriction"""
                   )
+
+options.register( 'sourceConnection'
+                , "oracle://cms_orcon_adg/CMS_RUNTIME_LOGGER"
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.string
+                , """beam data source connection string (aka PPS db)
+                     It's the source of crossing angle and beta * data"""
+                  )
+options.register( 'oms'
+                , "http://vocms0184.cern.ch/agg/api/v1"
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.string
+                , """OMS base URL"""
+                  )
+
+#duringFill mode specific:
+options.register( 'lastLumiFile'
+                , ''
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.string
+                , """duringFill only: path to file with lumiid to override the last lumisection processed by HLT.
+                     Used for testing. Leave empty for production behaviour (getting this info from OMS)"""
+                  )
+options.register( 'frontierKey'
+                , ''
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.string
+                , """duringFill only: run-unique key for writing with OnlinePopCon
+                     (used for confirming proper upload)"""
+                  )
+options.register('offsetLS'
+                , 2
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.int
+                , """duringFill only: offset between lastLumi (last LS processed by HLT or overriden by lastLumiFile)
+                     and the IOV of the payload to be uploaded"""
+                  )
+options.register( 'debugLogic'
+                , False
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.bool
+                , """duringFill only: Enables debug logic, meant to be used only for tests"""
+                  )
+options.register( 'defaultXangleX'
+                , 160.0 # urad
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.float
+                , """duringFill only: crossingAngleX value (in urad) for the default payload.
+                     The default payload is inserted after the last processed fill has ended
+                     and there's no ongoing stable beam yet. """
+                  )
+options.register( 'defaultXangleY'
+                , 0.0 # urad
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.float
+                , """duringFill only: crossingAngleY value (in urad) for the default payload.
+                     The default payload is inserted after the last processed fill has ended
+                     and there's no ongoing stable beam yet. """
+                  )
+options.register( 'defaultBetaX'
+                , 0.3 # meters
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.float
+                , """duringFill only: betaStarX value (in meters) for the default payload.
+                     The default payload is inserted after the last processed fill has ended
+                     and there's no ongoing stable beam yet. """
+                  )
+options.register( 'defaultBetaY'
+                , 0.3 # meters
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.float
+                , """duringFill only: betaStarY value (in meters) for the default payload.
+                     The default payload is inserted after the last processed fill has ended
+                     and there's no ongoing stable beam yet. """
+                  )
+
+
+# it's unlikely to ever use values different from the defaults, added as a parameter just in case
+options.register('minBetaStar',  0.1 
+                , VarParsing.VarParsing.multiplicity.singleton, VarParsing.VarParsing.varType.float
+                , """duringFill only: [meters] min value of the range of valid values.
+                     If the value is outside of this range the payload is not uploaded""")
+options.register('maxBetaStar',  100.
+                , VarParsing.VarParsing.multiplicity.singleton, VarParsing.VarParsing.varType.float
+                , """duringFill only: [meters] min value of the range of valid values.
+                     If the value is outside of this range the payload is not uploaded""")
+options.register('minCrossingAngle',  10.
+                , VarParsing.VarParsing.multiplicity.singleton, VarParsing.VarParsing.varType.float
+                , """duringFill only: [urad] min value of the range of valid values.
+                     If the value is outside of this range the payload is not uploaded""")
+options.register('maxCrossingAngle',  500.
+                , VarParsing.VarParsing.multiplicity.singleton, VarParsing.VarParsing.varType.float
+                , """duringFill only: [urad] min value of the range of valid values.
+                     If the value is outside of this range the payload is not uploaded""")
+
+# as the previous options, so far there was no need to use option, added just in case
+options.register( 'authenticationPath'
+                , ""
+                , VarParsing.VarParsing.multiplicity.singleton
+                , VarParsing.VarParsing.varType.string
+                , """for now this option was always left empty"""
+                  )
+
 options.parseArguments()
 if options.mode is None:
   raise ValueError("mode argument not provided. Supported modes are: duringFill endFill")
@@ -68,6 +167,7 @@ if options.mode not in ("duringFill", "endFill"):
 
 CondDBConnection = CondDB.clone( connect = cms.string( options.destinationConnection ) )
 CondDBConnection.DBParameters.messageLevel = cms.untracked.int32( options.messageLevel )
+CondDBConnection.DBParameters.authenticationPath = cms.untracked.string(options.authenticationPath)
 
 process.MessageLogger = cms.Service("MessageLogger",
                                     cout = cms.untracked.PSet(threshold = cms.untracked.string('INFO')),
@@ -87,29 +187,61 @@ if options.mode == 'endFill':
 else:
   timetype = 'lumiid'
 
-process.PoolDBOutputService = cms.Service("PoolDBOutputService",
-                                          CondDBConnection,
-                                          timetype = cms.untracked.string(timetype),
-                                          toPut = cms.VPSet(cms.PSet(record = cms.string('LHCInfoPerLSRcd'),
-                                                                     tag = cms.string( options.tag )
-                                                                     )
-                                                            )
-                                          )
+if options.mode == "endFill":
+  process.PoolDBOutputService = cms.Service("PoolDBOutputService",
+                                            CondDBConnection,
+                                            timetype = cms.untracked.string(timetype),
+                                            toPut = cms.VPSet(cms.PSet(record = cms.string('LHCInfoPerLSRcd'),
+                                                                      tag = cms.string( options.tag )
+                                                                      )
+                                                              )
+                                            )
+else:
+  process.OnlineDBOutputService = cms.Service("OnlineDBOutputService",
+    CondDBConnection,
+    preLoadConnectionString = cms.untracked.string('frontier://FrontierProd/CMS_CONDITIONS' 
+                                                   if not options.destinationConnection.startswith('sqlite') 
+                                                   else options.destinationConnection ),
+    lastLumiFile = cms.untracked.string(options.lastLumiFile),
+    omsServiceUrl = cms.untracked.string('http://cmsoms-eventing.cms:9949/urn:xdaq-application:lid=100/getRunAndLumiSection'
+                                         if not options.lastLumiFile else "" ),
+    # runNumber = cms.untracked.uint64(384468), #not used in production, the last LS processed is set as the 1st LS of this
+                                                #run if the omsServiceUrl is empty and file specified in lastLumiFile is empty
+    latency = cms.untracked.uint32(options.offsetLS),
+    timetype = cms.untracked.string(timetype),
+    toPut = cms.VPSet(cms.PSet(
+        record = cms.string('LHCInfoPerLSRcd'),
+        tag = cms.string( options.tag ),
+        onlyAppendUpdatePolicy = cms.untracked.bool(True)
+    )),
+    frontierKey = cms.untracked.string(options.frontierKey)
+)
 
-process.Test1 = cms.EDAnalyzer("LHCInfoPerLSPopConAnalyzer",
+
+process.Test1 = cms.EDAnalyzer(("LHCInfoPerLSPopConAnalyzer" if options.mode == "endFill" 
+                               else "LHCInfoPerLSOnlinePopConAnalyzer"),
                                SinceAppendMode = cms.bool(True),
                                record = cms.string('LHCInfoPerLSRcd'),
                                name = cms.untracked.string('LHCInfo'),
-                               Source = cms.PSet(fill = cms.untracked.uint32(6417),
+                               Source = cms.PSet(
                                    startTime = cms.untracked.string(options.startTime),
                                    endTime = cms.untracked.string(options.endTime),
-                                   endFill = cms.untracked.bool(True if options.mode == "endFill" else False),
+                                   endFill = cms.untracked.bool(options.mode == "endFill"),
                                    name = cms.untracked.string("LHCInfoPerLSPopConSourceHandler"),
-                                   connectionString = cms.untracked.string("oracle://cms_orcon_adg/CMS_RUNTIME_LOGGER"),
-                                   omsBaseUrl = cms.untracked.string("http://vocms0184.cern.ch/agg/api/v1"),
-                                   authenticationPath = cms.untracked.string(""),
-                                   debug=cms.untracked.bool(False)
-                                                 ),
+                                   connectionString = cms.untracked.string(options.sourceConnection),
+                                   omsBaseUrl = cms.untracked.string(options.oms),
+                                   authenticationPath = cms.untracked.string(options.authenticationPath),
+                                   debug=cms.untracked.bool(False), # Additional logs
+                                   debugLogic=cms.untracked.bool(options.debugLogic),
+                                   defaultCrossingAngleX = cms.untracked.double(options.defaultXangleX),
+                                   defaultCrossingAngleY = cms.untracked.double(options.defaultXangleY),
+                                   defaultBetaStarX = cms.untracked.double(options.defaultBetaX),
+                                   defaultBetaStarY = cms.untracked.double(options.defaultBetaY),
+                                   minBetaStar = cms.untracked.double(options.minBetaStar),
+                                   maxBetaStar = cms.untracked.double(options.maxBetaStar),
+                                   minCrossingAngle = cms.untracked.double(options.minCrossingAngle),
+                                   maxCrossingAngle = cms.untracked.double(options.maxCrossingAngle),
+                               ),
                                loggingOn = cms.untracked.bool(True),
                                IsDestDbCheckedInQueryLog = cms.untracked.bool(False)
                                )
