+++ CondTools/RunInfo/test/last_lumi.txt
@@ -1,0 +1,1 @@
+1663505258250241
